+++ colors/limei.lua
@@ -1,0 +1,1 @@
+require("limei").load()