+++ lua/limei/groups/plugins/init.lua
@@ -18,7 +18,7 @@ local modules = {
 function M.get(colors, options)
   local groups = {}
   for _, name in ipairs(modules) do
-    groups = vim.tbl_extend("force", groups, require("cold.groups.plugins." .. name).get(colors, options))
+    groups = vim.tbl_extend("force", groups, require("limei.groups.plugins." .. name).get(colors, options))
   end
   return groups
 end